Consolidate search utility functions
And remove them from the components
